requirements.txt
browser-use==0.7.9
python-dotenv==1.1.1
playwright==1.49.0